Add filtering logic to the treeplot nodes.
Make treeplot color tree nodes with a given list of commits.
Calculate the coloring using the ratio of commits which are included in the
filter file to the total number of commits that are represented by the tree
node.

Make FiltStats function fill the dictionary and keep track of the maximum
in-filter to total number of commits ratio.

Make the in GV_color_filt function use the maximum value to adjust the node
colors.